Refactor of functions on more friendly named modules
