+++ src/constants.js
@@ -1,0 +1,7 @@
+export const DEFAULT_CRITERIA = {
+  slaOptions: false,
+  selectedSla: true,
+  seller: true,
+  shippingEstimate: true,
+  deliveryChannel: true,
+}
