+++ src/package.js
@@ -1,0 +1,170 @@
+import { getShippingEstimateQuantityInSeconds } from '@vtex/estimate-calculator'
+
+import { getNewItems, getDeliveredItems } from './items'
+import { hydratePackageWithLogisticsExtraInfo } from './shipping'
+import { DEFAULT_CRITERIA } from './constants'
+
+export function groupPackages(items, criteria) {
+  return addToPackage(items, criteria, (packages, item) => {
+    return packages.find(pack => pack.package.index === item.package.index)
+  })
+}
+
+export function groupDeliveries(items, criteria) {
+  return addToPackage(items, criteria, (packages, item) => {
+    return packages.find(pack => {
+      if (
+        criteria.shippingEstimate &&
+        criteria.selectedSla &&
+        pack.shippingEstimate !== item.shippingEstimate
+      ) {
+        return false
+      }
+
+      if (criteria.slaOptions) {
+        const packSlas = pack.slas.reduce((acc, sla) => acc + sla.id, '')
+        const itemSlas = item.slas.reduce((acc, sla) => acc + sla.id, '')
+
+        if (packSlas !== itemSlas) {
+          return false
+        }
+      }
+
+      if (criteria.seller && pack.seller !== item.item.seller) {
+        return false
+      }
+
+      if (criteria.selectedSla && pack.selectedSla !== item.selectedSla) {
+        return false
+      }
+
+      const hasDeliveryWindow = pack.deliveryWindow && item.deliveryWindow
+
+      const areWindowsDifferent =
+        hasDeliveryWindow &&
+        pack.deliveryWindow.startDateUtc !== item.deliveryWindow.startDateUtc &&
+        pack.deliveryWindow.endDateUtc !== item.deliveryWindow.endDateUtc
+
+      if (
+        criteria.selectedSla &&
+        ((hasDeliveryWindow && areWindowsDifferent) ||
+          !!pack.deliveryWindow !== !!item.deliveryWindow)
+      ) {
+        return false
+      }
+
+      if (
+        criteria.deliveryChannel &&
+        pack.deliveryChannel !== item.deliveryChannel
+      ) {
+        return false
+      }
+
+      return true
+    })
+  })
+}
+
+export function addToPackage(items, criteria, fn) {
+  return items.reduce((packages, item) => {
+    const pack = fn(packages, item)
+    if (pack) {
+      if (
+        criteria.selectedSla &&
+        getShippingEstimateQuantityInSeconds(pack.shippingEstimate) <
+          getShippingEstimateQuantityInSeconds(item.shippingEstimate)
+      ) {
+        pack.shippingEstimate = item.shippingEstimate
+        pack.shippingEstimateDate = item.shippingEstimateDate
+      }
+
+      if (!criteria.selectedSla) {
+        pack.slas = pack.slas.concat(item.slas)
+      }
+
+      pack.items = pack.items.concat(item.item)
+      return packages
+    }
+
+    const newPackage = {
+      items: [item.item],
+      package: item.package,
+      slas: item.slas,
+      pickupFriendlyName: criteria.selectedSla
+        ? item.pickupFriendlyName
+        : undefined,
+      seller: criteria.seller ? item.item.seller : undefined,
+      address: criteria.selectedSla ? item.address : undefined,
+      selectedSla: criteria.selectedSla ? item.selectedSla : undefined,
+      deliveryIds: item.deliveryIds,
+      deliveryChannel: criteria.deliveryChannel
+        ? item.deliveryChannel
+        : undefined,
+      deliveryWindow: criteria.selectedSla ? item.deliveryWindow : undefined,
+      shippingEstimate: criteria.selectedSla
+        ? item.shippingEstimate
+        : undefined,
+      shippingEstimateDate: criteria.selectedSla
+        ? item.shippingEstimateDate
+        : undefined,
+      item: undefined,
+    }
+
+    return packages.concat(newPackage)
+  }, [])
+}
+
+export function parcelify(order, options = {}) {
+  const {
+    items = [],
+    packageAttachment = {},
+    shippingData = {},
+    changesAttachment = {},
+  } = order
+
+  const criteria = {
+    ...DEFAULT_CRITERIA,
+    ...(options.criteria ? options.criteria : {}),
+  }
+
+  const packages =
+    packageAttachment && packageAttachment.packages
+      ? packageAttachment.packages
+      : []
+  const logisticsInfo =
+    shippingData && shippingData.logisticsInfo ? shippingData.logisticsInfo : []
+  const selectedAddresses =
+    shippingData && shippingData.selectedAddresses
+      ? shippingData.selectedAddresses
+      : []
+  const changes =
+    changesAttachment && changesAttachment.changesData
+      ? changesAttachment.changesData
+      : []
+
+  const itemsWithIndex = items.map((item, index) => ({ ...item, index }))
+  const packagesWithIndex = packages.map((pack, index) => ({ ...pack, index }))
+
+  const itemsCleaned = getNewItems(itemsWithIndex, changes)
+
+  const deliveredItems = getDeliveredItems({
+    items: itemsCleaned,
+    packages: packagesWithIndex,
+  })
+
+  const enhancePackage = pkg =>
+    hydratePackageWithLogisticsExtraInfo(pkg, logisticsInfo, selectedAddresses)
+
+  deliveredItems.delivered = deliveredItems.delivered.map(enhancePackage)
+  deliveredItems.toBeDelivered = deliveredItems.toBeDelivered.map(
+    enhancePackage
+  )
+
+  const deliveredPackages = groupPackages(deliveredItems.delivered, criteria)
+  const toBeDeliveredPackages = groupDeliveries(
+    deliveredItems.toBeDelivered,
+    criteria
+  )
+
+  return deliveredPackages.concat(toBeDeliveredPackages)
+}
